3.5
falta playlists
e fazer a relacao entre todas

+++ routes/web.php
@@ -33,28 +33,34 @@ Route::get('/homepage',[HomeController::class,'homepage'])->name('home');
 
 Auth::routes();
 
-//->name('admin.home')->middleware('is_admin')
 
+Route::get('/users', [UserController::class,'index'])->name('users.index')->middleware('is_admin');
+Route::get('/users/create', [UserController::class,'create'])->name('users.create')->middleware('is_admin');
+Route::post('/users', [UserController::class,'store'])->name('users.store')->middleware('is_admin');
+Route::get('/users/{user}', [UserController::class,'show'])->name('users.show')->middleware('is_admin');
+Route::get('/users/{user}/edit', [UserController::class,'edit'])->name('users.edit')->middleware('is_admin');
+Route::put('/users/{user}', [UserController::class,'update'])->name('users.update')->middleware('is_admin');
+Route::delete('/users/{user}', [UserController::class,'destroy'])->name('users.destroy')->middleware('is_admin');
 
 
 
 Route::get('/artists', [ArtistController::class,'index'])->name('artists.index');
-Route::get('/artists/create', [ArtistController::class,'create'])->name('artists.create');
-Route::post('/artists', [ArtistController::class,'store'])->name('artists.store');
+Route::get('/artists/create', [ArtistController::class,'create'])->name('artists.create')->middleware('is_admin');
+Route::post('/artists', [ArtistController::class,'store'])->name('artists.store')->middleware('is_admin');
 Route::get('/artists/{artist}', [ArtistController::class,'show'])->name('artists.show');
-Route::get('/artists/{artist}/edit', [ArtistController::class,'edit'])->name('artists.edit');
-Route::put('/artists/{artist}', [ArtistController::class,'update'])->name('artists.update');
-Route::delete('/artists/{artist}', [ArtistController::class,'destroy'])->name('artists.destroy');
+Route::get('/artists/{artist}/edit', [ArtistController::class,'edit'])->name('artists.edit')->middleware('is_admin');
+Route::put('/artists/{artist}', [ArtistController::class,'update'])->name('artists.update')->middleware('is_admin');
+Route::delete('/artists/{artist}', [ArtistController::class,'destroy'])->name('artists.destroy')->middleware('is_admin');
 
 
 
 Route::get('/musicas', [MusicaController::class,'index'])->name('musicas.index');
-Route::get('/musicas/create', [MusicaController::class,'create'])->name('musicas.create');
-Route::post('/musicas', [MusicaController::class,'store'])->name('musicas.store');
+Route::get('/musicas/create', [MusicaController::class,'create'])->name('musicas.create')->middleware('is_admin');
+Route::post('/musicas', [MusicaController::class,'store'])->name('musicas.store')->middleware('is_admin');
 Route::get('/musicas/{musica}', [MusicaController::class,'show'])->name('musicas.show');
-Route::get('/musicas/{musica}/edit', [MusicaController::class,'edit'])->name('musicas.edit');
-Route::put('/musicas/{musica}', [MusicaController::class,'update'])->name('musicas.update');
-Route::delete('/musicas/{musica}', [MusicaController::class,'destroy'])->name('musicas.destroy');
+Route::get('/musicas/{musica}/edit', [MusicaController::class,'edit'])->name('musicas.edit')->middleware('is_admin');
+Route::put('/musicas/{musica}', [MusicaController::class,'update'])->name('musicas.update')->middleware('is_admin');
+Route::delete('/musicas/{musica}', [MusicaController::class,'destroy'])->name('musicas.destroy')->middleware('is_admin');
 
 
 
@@ -69,20 +75,20 @@ Route::delete('/playlists/{playlist}', [PlaylistController::class,'destroy'])->n
 
 
 Route::get('/albuns', [AlbumController::class,'index'])->name('albuns.index');
-Route::get('/albuns/create', [AlbumController::class,'create'])->name('albuns.create');
-Route::post('/albuns', [AlbumController::class,'store'])->name('albuns.store');
+Route::get('/albuns/create', [AlbumController::class,'create'])->name('albuns.create')->middleware('is_admin');
+Route::post('/albuns', [AlbumController::class,'store'])->name('albuns.store')->middleware('is_admin');
 Route::get('/albuns/{album}', [AlbumController::class,'show'])->name('albuns.show');
-Route::get('/albuns/{album}/edit', [AlbumController::class,'edit'])->name('albuns.edit');
-Route::put('/albuns/{album}', [AlbumController::class,'update'])->name('albuns.update');
-Route::delete('/albuns/{album}', [AlbumController::class,'destroy'])->name('albuns.destroy');
+Route::get('/albuns/{album}/edit', [AlbumController::class,'edit'])->name('albuns.edit')->middleware('is_admin');
+Route::put('/albuns/{album}', [AlbumController::class,'update'])->name('albuns.update')->middleware('is_admin');
+Route::delete('/albuns/{album}', [AlbumController::class,'destroy'])->name('albuns.destroy')->middleware('is_admin');
 
 
 
 Route::get('/generos', 'App\Http\Controllers\GeneroController@index')->name('generos.index');
-Route::get('/generos/create', 'App\Http\Controllers\GeneroController@create')->name('generos.create');
-Route::post('/generos', 'App\Http\Controllers\GeneroController@store')->name('generos.store');
+Route::get('/generos/create', 'App\Http\Controllers\GeneroController@create')->name('generos.create')->middleware('is_admin');
+Route::post('/generos', 'App\Http\Controllers\GeneroController@store')->name('generos.store')->middleware('is_admin');
 Route::get('/generos/{genero}', 'App\Http\Controllers\GeneroController@show')->name('generos.show');
-Route::get('/generos/{genero}/edit', 'App\Http\Controllers\GeneroController@edit')->name('generos.edit');
-Route::put('/generos/{genero}', 'App\Http\Controllers\GeneroController@update')->name('generos.update');
-Route::delete('/generos/{genero}', 'App\Http\Controllers\GeneroController@destroy')->name('generos.destroy');
+Route::get('/generos/{genero}/edit', 'App\Http\Controllers\GeneroController@edit')->name('generos.edit')->middleware('is_admin');
+Route::put('/generos/{genero}', 'App\Http\Controllers\GeneroController@update')->name('generos.update')->middleware('is_admin');
+Route::delete('/generos/{genero}', 'App\Http\Controllers\GeneroController@destroy')->name('generos.destroy')->middleware('is_admin');
